requirements.txt
fastapi
uvicorn
jinja2
tabula-py
pandas
openpyxl
python-multipart